@@ -106,11 +106,11 @@
                     <td>@Html.DisplayFor(modelItem => item.EmployeeType)</td>
                     <td>@Html.DisplayFor(modelItem => item.IsActive)</td>
                     <td>@Html.DisplayFor(modelItem => item.CreationDate)</td>
-                <td class="d-flex justify-content-around">
-                    <a class="btn btn-info" asp-action="Details" asp-route-id="@item.Id"><i class="fas fa-eye"></i>Details</a> |
-                    <a class="btn btn-success" asp-action="Edit" asp-route-id="@item.Id"><i class="fas fa-edit"></i>Edit</a> |
-                    <a class="btn btn-danger" asp-action="Delete" asp-route-id="@item.Id"><i class="fas fa-trash-alt"></i>Delete</a>
-                </td>
+                    <td class="d-flex justify-content-around">
+                        <a class="btn btn-info" asp-action="Details" asp-route-id="@item.Id"><i class="fas fa-eye"></i>Details</a> |
+                        <a class="btn btn-success" asp-action="Edit" asp-route-id="@item.Id"><i class="fas fa-edit"></i>Edit</a> |
+                        <a class="btn btn-danger" asp-action="Delete" asp-route-id="@item.Id"><i class="fas fa-trash-alt"></i>Delete</a>
+                    </td>
                 </tr>
             }
             }
